backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Bruno Amaral Teixeira de Freitas <brunoatf257@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
strawberry-graphql = "^0.235.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.1"
pylint = "^3.2.4"
pytest = "^8.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
